reorganize handlers
