+++ api/server.go
@@ -3,14 +3,11 @@ package server
 
 import (
 	"crypto/tls"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"log"
 	"net/http"
 	"os"
-	"strconv"
-	"strings"
 	"time"
 
 	"github.com/canonical/gocert/internal/certdb"
@@ -88,23 +85,14 @@ func NewServer(configFile string) (*http.Server, error) {
 
 	env := &environment{}
 	env.db = db
-	router := http.NewServeMux()
-	router.HandleFunc("GET /certificate_requests", GetCertificateRequests(env))
-	router.HandleFunc("POST /certificate_requests", PostCertificateRequests(env))
-	router.HandleFunc("GET /certificate_requests/{id}", GetCertificateRequest(env))
-	router.HandleFunc("DELETE /certificate_requests/{id}", DeleteCertificateRequest(env))
-	router.HandleFunc("POST /certificate_requests/{id}/certificate", PostCertificate(env))
-
-	v1 := http.NewServeMux()
-	v1.HandleFunc("GET /status", HealthCheck)
-	v1.Handle("/api/v1/", http.StripPrefix("/api/v1", router))
+	router := GoCertRouter(env)
 
 	s := &http.Server{
 		Addr: fmt.Sprintf(":%d", config.Port),
 
 		ReadTimeout:    10 * time.Second,
 		WriteTimeout:   10 * time.Second,
-		Handler:        Logging(v1),
+		Handler:        router,
 		MaxHeaderBytes: 1 << 20,
 		TLSConfig: &tls.Config{
 			Certificates: []tls.Certificate{serverCerts},
@@ -112,117 +100,4 @@ func NewServer(configFile string) (*http.Server, error) {
 	}
 
 	return s, nil
-}
-
-func HealthCheck(w http.ResponseWriter, r *http.Request) {
-	w.Write([]byte("Server Alive")) //nolint:errcheck
-}
-
-func Logging(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		next.ServeHTTP(w, r)
-		log.Println(r.Method, r.URL.Path)
-	})
-}
-
-func GetCertificateRequests(env *environment) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		certs, err := env.db.RetrieveAll()
-		if err != nil {
-			logError(err.Error(), 500, w)
-			return
-		}
-		body, err := json.Marshal(certs)
-		if err != nil {
-			logError(err.Error(), 500, w)
-			return
-		}
-		w.Write(body)
-	}
-}
-
-func PostCertificateRequests(env *environment) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		csr := make([]byte, r.ContentLength)
-		bytesRead, err := r.Body.Read(csr)
-		if bytesRead != int(r.ContentLength) {
-			logError("couldn't read the body completely", 400, w)
-			return
-		}
-		if err.Error() != "EOF" {
-			logError(err.Error(), 500, w)
-			return
-		}
-		id, err := env.db.Create(string(csr))
-		if err != nil {
-			if strings.Contains(err.Error(), "UNIQUE constraint failed") {
-				logError("given csr already recorded", 400, w)
-				return
-			} else {
-				logError(err.Error(), 500, w)
-				return
-			}
-		}
-		w.WriteHeader(201)
-		w.Write([]byte(strconv.FormatInt(id, 10)))
-	}
-}
-
-func GetCertificateRequest(env *environment) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		id := r.PathValue("id")
-		cert, err := env.db.Retrieve(id)
-		if err != nil {
-			logError(err.Error(), 500, w)
-			return
-		}
-		body, err := json.Marshal(cert)
-		if err != nil {
-			logError(err.Error(), 500, w)
-			return
-		}
-		w.Write(body)
-	}
-}
-
-func DeleteCertificateRequest(env *environment) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		id := r.PathValue("id")
-		err := env.db.Delete(id)
-		if err != nil {
-			logError(err.Error(), 500, w)
-			return
-		}
-		w.WriteHeader(204)
-	}
-}
-
-func PostCertificate(env *environment) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		cert := make([]byte, r.ContentLength)
-		bytesRead, err := r.Body.Read(cert)
-		if bytesRead != int(r.ContentLength) {
-			logError("couldn't read the body completely", 400, w)
-			return
-		}
-		if err.Error() != "EOF" {
-			logError(err.Error(), 500, w)
-			return
-		}
-		id := r.PathValue("id")
-		insertId, err := env.db.Update(id, string(cert))
-		if err != nil {
-			logError(err.Error(), 500, w)
-			return
-		}
-		w.WriteHeader(201)
-		w.Write([]byte(strconv.FormatInt(insertId, 10)))
-	}
-}
-
-func logError(msg string, status int, w http.ResponseWriter) {
-	errMsg := fmt.Sprintf("error: %s", msg)
-	log.Println(errMsg)
-	w.WriteHeader(status)
-	w.Write([]byte(errMsg))
 }
